+++ include/utils.h
@@ -4,7 +4,7 @@
 #include <vector>
 using std::vector;
 
-double calculate_distance(float normSize, vector<float>&,vector<float>&);
+double calculate_distance(unsigned int normSize, vector<float>&,vector<float>&);
 double vector_multiply(vector<float>&,vector<float>&);
 
 #endif
